Removed while loop, opted for cron job instead.  Also added /r/hockey as a subreddit to monitor

+++ app/logger.py
@@ -5,9 +5,9 @@ import os
 import time
  
 class Logger(object):
-  def __init__(self, filename="log.txt"):
+  def __init__(self, filename="log.txt", newFile=False):
     self.filename = filename
-    if os.path.isfile(self.filename):
+    if os.path.isfile(self.filename) and newFile:
       os.remove(self.filename)
       
     self.write("Starting on "+time.strftime("%Y-%m-%d %H:%M:%S"))
